✨ Added rooms and cleaned up repo/single Dockerfile
