--- canvas-game/.gitignore
@@ -1,6 +1,0 @@
-.cache
-dist
-node_modules
-.idea/
-src/
-package-lock.json
