+++ db/ethereum/testnet/node/CURRENT
@@ -1,0 +1,1 @@
+MANIFEST-000002
