+++ db/ethereum/testnet/node/LOG
@@ -1,0 +1,1 @@
+2020/08/12-14:46:42.433325 70000b503000 Delete type=3 #1
