+++ docker-compose.yaml
@@ -10,14 +10,14 @@ services:
     networks:
       - envoy
     environment:
-      - BOOTSTRAP_SERVERS=/ip4/34.65.237.196/tcp/9091/p2p/16Uiu2HAm1pyd27aNRqx7icG5uDBKNFDwvwaEg2vuAUEtLN1pRa5X,/ip4/34.65.139.63/tcp/9091/p2p/16Uiu2HAmR2va1xavPsYRrpCVnjvcUr1Jqjf9UjywdyPxeZ5Z3eLZ
+      - BOOTSTRAP_SERVERS=/ip4/34.65.75.45/tcp/9091/p2p/16Uiu2HAm2cjqsDMmprtN2QKaq3LJrq3YK7vtdbQQFsxGrhLRoYsy,/ip4/34.65.177.154/tcp/9091/p2p/16Uiu2HAm9C4oJPeRkdXnYxtXzFpDqpcCbWLsgNW4irrCLZTJ7cBd
     volumes:
       - /app/db
 
   client:
     image: gcr.io/hoprassociation/hopr-mazebot/client
     build:
-      context: ./canvas-game
+      context: .
     ports:
       - 3000:3000
     networks:
