+++ index.html
@@ -88,6 +88,10 @@
             <div class='status-label'>Position</div>
             <div id='current-position' class='status-value'></div>
           </div>
+          <div>
+            <div class='status-label'>Room</div>
+            <div id='current-room' class='status-value'></div>
+          </div>
         </div>
       </div>
