+++ js/display/avatar-path-renderer.js
