+++ js/display/status-bar.js
@@ -4,7 +4,8 @@ function StatusBar() {
   return {
     setMaze: setMaze,
     setPosition: setPosition,
-    setMoveCount: setMoveCount
+    setMoveCount: setMoveCount,
+    setRoom: setRoom
   }
 
   function setMaze(maze) {
@@ -19,5 +20,9 @@ function StatusBar() {
 
   function setPosition(position) {
     document.getElementById('current-position').innerText = position[0] + ', ' + position[1];
+  }
+
+  function setRoom(room) {
+    document.getElementById('current-room').innerText = room;
   }
 }
