+++ process.yaml
@@ -1,0 +1,12 @@
+apps:
+    - script: ./server.js
+      name: "server"
+      instances: 1
+    - script: ./envoy/envoy.sh
+      name: "proxy"
+      instances: 1
+      env:
+        SERVICE_ADDRESS: 127.0.0.1
+    - script: npm start
+      name: "client"
+      instances: 1
