+++ ts/server.ts
@@ -1,0 +1,2 @@
+require('@hoprnet/hopr-server')
+console.log(`Starting HOPR Server`);
